rename to generic_passband
